requirements.txt
streamlit
easyocr
Pillow
sympy